Change float to int16_t

+++ app/src/main/cpp/shared/DefaultAudioStreamCallback.h
@@ -45,7 +45,7 @@ public:
             mIsThreadAffinitySet = true;
         }
 
-        float *outputBuffer = static_cast<float *>(audioData);
+        int16_t *outputBuffer = static_cast<int16_t *>(audioData);
 
         std::shared_ptr<IRenderableAudio> localRenderable = mRenderable;
         if (!localRenderable) {
